extract pm quality fairness routes
